% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_canton_pollen_forecast.R
\name{get_canton_pollen_forecast}
\alias{get_canton_pollen_forecast}
\title{get_canton_pollen_forecast returns a forecast of the pollen distribution of
eacch canton in a dataframe form}
\usage{
get_canton_pollen_forecast()
}
\value{
a dataframe containing pollen forecast per canton
}
\description{
get_canton_pollen_forecast returns a forecast of the pollen distribution of
eacch canton in a dataframe form
}
\examples{
get_canton_pollen_forecast() -> df
}
