% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_helpers.R
\name{get_map_path}
\alias{get_map_path}
\title{Get full path to a shapefile in extdata/maps/}
\usage{
get_map_path(type = c("see", "kanton", "berggebiete"), filename = NULL)
}
\arguments{
\item{type}{The name of the map folder, e.g. "see", "kanton", "berggebiete"}

\item{filename}{The shapefile name (default for standard naming)}
}
\value{
Full path to the requested shapefile
}
\description{
Get full path to a shapefile in extdata/maps/
}
\examples{
get_map_path("see")
get_map_path("kanton")
get_map_path("berggebiete")
}
