% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_pollen_forecast}
\alias{get_pollen_forecast}
\title{get_pollen_forecast: fetches information from api and loads the data.}
\usage{
get_pollen_forecast()
}
\value{
a dataframe containing all information about pollen forecast.
}
\description{
get_pollen_forecast: fetches information from api and loads the data.
}
\examples{
get_pollen_forecast() -> pollen_df
}
