% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_swiss_cities_path.R
\name{get_swiss_cities_path}
\alias{get_swiss_cities_path}
\title{Get full path to SwissCities.csv}
\usage{
get_swiss_cities_path()
}
\value{
A string with the full path to SwissCities.csv
}
\description{
This function returns the full path to the SwissCities.csv file
located in inst/extdata/.
}
\examples{
swiss_cities <- read.csv(path)
}
