% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_cantons.R
\name{plot_cantons}
\alias{plot_cantons}
\title{plot_cantons is a function that plots cantons data}
\usage{
plot_cantons(
  canton_path,
  lake_path = NULL,
  value_df,
  map_crs = 2056,
  value_colname = "avg",
  title = "Map"
)
}
\arguments{
\item{canton_path}{a string value, defines path to canton data}

\item{lake_path}{a string value, defines path to lake data}

\item{value_df}{a tibble, determines the dataframeof values to plot}

\item{map_crs}{an integer value, determines the reference system of the
geographical data. Basic value is 2056.}

\item{value_colname}{a string value, determines the name of the column plotted}

\item{title}{a string value, determines the overall title of the plot}
}
\value{
Nothing, directly plots a map of cantons
}
\description{
plot_cantons is a function that plots cantons data
}
\examples{
plot_cantons(canton_path = "path_to_cantons",
                      lake_path = "path_to_lake",
                      value_df = df,
                      value_colname = "avg",
                      title = "Map of Cantons") -> NONE
}
