% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{plot_map}
\alias{plot_map}
\title{Title}
\usage{
plot_map(
  canton_type = "kanton",
  mountain_type = "berggebiete",
  lake_type = NULL,
  points_df = NULL,
  points_crs = 4326,
  map_crs = 2056,
  base_fill = rgb(176/256, 227/256, 170/256),
  mountain_fill = "antiquewhite2",
  lake_fill = "lightblue",
  border_color = "black",
  point_color = "red",
  point_size = 2,
  show_labels = TRUE,
  title = "Map"
)
}
\arguments{
\item{points_df}{a tibble or dataframe, contains a list of all points to be
plotted on the graph. Basic value is NONE.}

\item{points_crs}{an integer value, crs stands for coordinate reference system,
specifies latitude and longitude in degrees of a given point.
Basic value is 4326, same as standard GPS value.}

\item{map_crs}{an integer value, crs stands for coordinate reference system,
specifies the crs for countries and larger areas.
Basic value is 2056.}

\item{base_fill}{a string value, the string representing the colour of the
base map on the graph}

\item{mountain_fill}{a string value, the string representing the colour of
the mountain representation on the graph}

\item{lake_fill}{a string value, the string representing the colour of the
lake representation on the graph}

\item{border_color}{a string value, the string representing the colour of the
graph border}

\item{point_color}{a string value, the string representing the colour of the
points in the graph}

\item{point_size}{an integer value, size of the points on the graph}

\item{show_labels}{a boolean value, if to show the labels of the graph}

\item{title}{the title of the graph}

\item{canton_path}{the path to the file containing canton data}

\item{mountain_path}{the path to the file containing mountain data}

\item{lake_path}{the path to the file containing lake data}
}
\value{
Nothing, plots graph directly
}
\description{
Title
}
